Added Off Road Reset and Lap Time
Tagged the terrain outside of the road 'Grass' and made a function that reset the car to it's starting position if it stays on the grass for too long. Also quickly workshopped a lap function with working Stop watch and Best Time function. Could add checkpoints to have the player properly drive on the track and also have the finish line only function if the car passes from one side.

Outside of that, I want to make the car tank.